LPS-60972 Isolate gradle-plugins-test-integration configuration
